+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ## Enhancements
 - Fix lower bounds of dependency versions.
+- Allow to load multi-objective SMAC3v2 and add example (#69)
+- Do not disable existing loggers.
 - Update author email.
 - Add exit button which first deletes running jobs and then terminates DeepCave.
 - Nicer handling of Keyboard Interrupt.
@@ -15,6 +17,7 @@
 - Fix api examples (#68).
 - Reset inputs to fix error when subsequently selecting runs with different configspaces, objectives or budgets (#106).
 - Fix errors due to changing inputs before runselection (#64).
+- For fANOVA, remove constant hyperparameters from configspace (#9).
 
 # Version 1.1.3